+++ src/Flyback.Core/Graph/NodeCatalog.Patterns.cs
@@ -10,13 +10,13 @@ public partial class NodeCatalog
     {
         yield return new NodeDef(
             NoiseTypeId, "Noise", ModuleCategories.Patterns,
-            [..Position(), Num("z"), Num("scale", 2f, 0f, 32f)], [Num("out")],
+            [..Position(), Num("z"), Num("scale", 2f, 0f, 32f)], [Num("out", 0f, 0f, 1f)],
             (em, i) => [em.Ternary(OpCode.Noise3, em.Mul(i[0], i[3]), em.Mul(i[1], i[3]), i[2])],
             "Smooth random field in 0..1. Drive z from Time to make it boil.");
 
         yield return new NodeDef(
             "pattern.checker", "Checker", ModuleCategories.Patterns,
-            [..Position(), Num("size", 4f, 0f, 32f)], [Num("out")],
+            [..Position(), Num("size", 4f, 0f, 32f)], [Num("out", 0f, 0f, 1f)],
             (em, i) =>
             {
                 var fx = em.Unary(OpCode.Floor, em.Mul(i[0], i[2]));
@@ -27,7 +27,7 @@ public partial class NodeCatalog
 
         yield return new NodeDef(
             "pattern.rings", "Rings", ModuleCategories.Patterns,
-            [..Position(), Num("freq", 4f, 0f, 32f), Num("offset", 0f, 0f, 1f)], [Num("out")],
+            [..Position(), Num("freq", 4f, 0f, 32f), Num("offset", 0f, 0f, 1f)], [Num("out", 0f, -1f, 1f)],
             (em, i) =>
             {
                 var radius = em.Binary(OpCode.Hypot, i[0], i[1]);